pyproject.toml
[tool.poetry]
name = "lm_inductive_bias"
version = "0.1.0"
description = "Research on learnability on information-theoretic continuum"
authors = ["agiats"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
scipy = ">=1.15.1"
matplotlib = ">=3.10.0"
pandas = ">=2.2.3"
numpy = ">=1.26.4"
torch = ">=2.6.0"
rau = { git = "ssh://git@github.com/bdusell/rau.git", branch = "differentiable-stacks" }
seaborn = ">=0.13.2"
nltk = ">=3.9.1"
hydra-core = ">=1.3.2"
spacy = ">=3.8.4"
omegaconf = "^2.3.0"

[tool.poetry.group.dev.dependencies]
ruff = ">=0.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
# Enable pycodestyle, pyflakes, isort, and additional rules
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "N",  # pep8-naming
    "UP", # pyupgrade
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = [
    "E501", # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
